% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{B_basis}
\alias{B_basis}
\title{Spline Basis Functions}
\usage{
B_basis(xgrid, knots)
}
\arguments{
\item{xgrid}{grid where we want to evaluate the spline functions (vector of length n)}

\item{knots}{vector of knots for the splines (vector of length K)}
}
\value{
n x (K+1) - matrix representing the value of each basis function
}
\description{
Construct the J basis functions for the splines evaluated on a grid.
}
