% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{compute_WAIC}
\alias{compute_WAIC}
\title{Calculate WAIC}
\usage{
compute_WAIC(model_fit)
}
\arguments{
\item{model_fit}{results of a model fit from the lddmm function}
}
\value{
A scalar indicating the WAIC (smaller WAIC denotes better fit)
}
\description{
Function to compute the Watanabe-Akaike information criterion (Gelman, Hwang,
Vehtari, 2014), which estimates the expected out-of-sample-prediction error
using a bias-corrected adjustment of within-sample error.
}
