% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Example dataset}
\format{
A data frame with n rows and 6 variables
}
\usage{
data
}
\description{
A toy dataset in the correct format for the LDDMM function call. This dataset
has two possible response categories.
}
\details{
\itemize{
\item \code{subject}: vector of size n containing the participant labels
\item \code{block}: vector of size n containing the training blocks (longitudinal units)
\item \code{s}: vector of size n containing the stimuli
\item \code{d}: vector of size n containing the decisions
\item \code{r_time}: vector of size n containing the response times
\item \code{cens}: vector of size n containing the censoring indicators (1 censored, 0 non censored)
}
}
\keyword{datasets}
