% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{H_ball}
\alias{H_ball}
\title{Hamming Ball}
\usage{
H_ball(x, S, r)
}
\arguments{
\item{x}{center of the Hamming Ball}

\item{S}{number of states}

\item{r}{radius of the Hamming Ball}
}
\value{
Hamming Ball
}
\description{
Computes the Hamming Ball centered at x with radius r.
}
