% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{P_smooth1}
\alias{P_smooth1}
\title{Spline Penalty Matrix}
\usage{
P_smooth1(K)
}
\arguments{
\item{K}{Number of spline knots}
}
\value{
Covariance of the smoothness inducing prior (penalizing first
differences in the spline coefficients)
}
\description{
Construct the covariance matrix P of the smoothness inducing prior for the
spline coefficients
}
