% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{plot_accuracy}
\alias{plot_accuracy}
\title{Descriptive plots}
\usage{
plot_accuracy(data)
}
\arguments{
\item{data}{dataframe with the following columns:
\itemize{
\item subject: vector of size n containing the participant labels
\item block: vector of size n containing the training blocks (longitudinal units)
\item s: vector of size n containing the stimuli
\item d: vector of size n containing the decisions
\item r_time: vector of size n containing the log transformed of the response times (in milliseconds)
\item cens: vector of size n containing the censoring indicators (1 censored, 0 non censored)
}}
}
\value{
Individual and population level raw accuracies
}
\description{
Plot the accuracy of the raw data.
}
