% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{plot_post_pars}
\alias{plot_post_pars}
\title{Plot posterior estimates}
\usage{
plot_post_pars(data, fit, par = c("drift", "boundary"))
}
\arguments{
\item{data}{dataframe with the following columns:
\itemize{
\item subject: vector of size n containing the participant labels
\item block: vector of size n containing the training blocks (longitudinal units)
\item s: vector of size n containing the stimuli
\item d: vector of size n containing the decisions
\item r_time: vector of size n containing the response times
\item cens: vector of size n containing the censoring indicators (1 censored, 0 non censored)
}}

\item{fit}{fit from the lddmm function}

\item{par}{parameter to output ('drift', or 'boundary')}
}
\value{
Posterior mean and 95\% CI
}
\description{
Function to plot the posterior mean and credible intervals of the parameters of interest from a lddmm fit object.
}
