% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{plot_RT}
\alias{plot_RT}
\title{Descriptive plots}
\usage{
plot_RT(data)
}
\arguments{
\item{data}{dataframe with the following columns:
\itemize{
\item subject: vector of size n containing the participant labels
\item block: vector of size n containing the training blocks (longitudinal units)
\item s: vector of size n containing the stimuli
\item d: vector of size n containing the decisions
\item r_time: vector of size n containing the response times
\item cens: vector of size n containing the censoring indicators (1 censored, 0 non censored)
}}
}
\value{
population level raw response times
}
\description{
Plot the mean response times of the raw data.
}
